+++ ui_tetris/src/constants.rs
@@ -1,8 +1,8 @@
 use render_tetris::SCALE_FACTOR;
 
 // window
-pub const DEFAULT_WINDOW_WIDTH: u32 = (650.0 * SCALE_FACTOR) as u32;
-pub const DEFAULT_WINDOW_HEIGHT: u32 = (800.0 * SCALE_FACTOR) as u32;
+pub const DEFAULT_WINDOW_WIDTH: f64 = 650.0 * SCALE_FACTOR;
+pub const DEFAULT_WINDOW_HEIGHT: f64 = 800.0 * SCALE_FACTOR;
 
 // text
 pub const DEFAULT_FONT_SIZE: u32 = (16.0 * SCALE_FACTOR) as u32;
